+++ CHANGELOG.md
@@ -2,11 +2,13 @@
 ## 1.0.1
  ## 修复
  + UrlUtils 参数判断
+ + cache 序列化与反序列化失败问题
  ## 新增
   + ObjectUtils
   + Exceptions
   + DateUtils#between等方法
   + BeanUtils#getGetterMethods,BeanUtils#getSetterMethods等方法
+  + JedisPool Cache
  ## 删减
   + commons-cache 移除对hutool的依赖
 ## 1.0.0